pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

# ----------------------------------------------------------------------
# Project metadata (PEP 621)
# ----------------------------------------------------------------------
[project]
name            = "cost-model"
version         = "0.1.0"
description     = "Retirement-plan cost simulation"
readme          = "README.md"
license         = { text = "MIT" }
requires-python = ">=3.11"

classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]

# ----------------------------------------------------------------------
# Runtime dependencies
# ----------------------------------------------------------------------
dependencies = [
  "pandas>=2.2,<3.0",
  "numpy>=1.26,<2.0",
  "scipy>=1.13,<2.0",
  "scikit-learn>=1.4,<2.0",
  "lifelines>=0.28,<0.29",
  "matplotlib>=3.8,<4.0",
  "pyarrow>=15.0,<16.0",
  "lightgbm>=4.3,<5.0",
  "joblib>=1.4,<2.0",
  "pyyaml>=6.0,<7.0"
]

# ----------------------------------------------------------------------
# Optional / dev-only extras
# ----------------------------------------------------------------------
[project.optional-dependencies]
dev = [
  "mypy>=1.10",
  "flake8>=7.0",
  "black>=24.3",
  "isort>=5.13",
  "pytest>=8.0",
  "pytest-cov>=5.0"
]

# ----------------------------------------------------------------------
# Console scripts
# ----------------------------------------------------------------------
[project.scripts]
cost-model = "cost_model.projections.cli:main"

# ----------------------------------------------------------------------
# Setuptools package discovery
# ----------------------------------------------------------------------
[tool.setuptools.packages.find]
where   = ["."]
include = ["cost_model*"]

# Include typing marker file
[tool.setuptools.package-data]
"cost_model" = ["py.typed"]

# ----------------------------------------------------------------------
# Static-analysis / style tools
# ----------------------------------------------------------------------
[tool.mypy]
python_version        = "3.11"
strict                = true
ignore_missing_imports = true
show_error_codes      = true
pretty                = true

[tool.flake8]
max-line-length = 100
extend-ignore   = ["E203", "W503"]   # compatible with Black
exclude         = ["build", "dist", ".venv"]

[tool.black]
line-length     = 100
target-version  = ["py311"]

[tool.isort]
profile         = "black"
line_length     = 100
known_first_party = ["cost_model"]